refactor: centralize startup lifecycle runtime
